Extract User and Transaction table into separate files
